# documentation/API_specification.yaml
swagger: "2.0"
info:
  description: "The goal of the component consists in filtering blueprints and raking them, based on the application developer requirements. Source code in [github](https://github.com/DITAS-Project/data-utility-resolution-engine)"
  version: "0.0.1"
  title: "Data Utility Resolution Engine"
  license:
    name: "Apache License 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "TBD"
basePath: "/"
schemes:
- "http"
paths:
  /v1/filterBlueprints:
    post:
      summary: "Ranks and filters abstract blueprints based on application requirements"
      description: "The only API of this method filters a list of blueprints (input) that are complaint with the application requirements (input), and ranks them."
      consumes:
      - "application/json; charset=utf-8"
      produces:
      - "application/json; charset=utf-8"
      parameters:
      - in: "body"
        name: "Input"
        description: "Application requirements and list of VDC blueprints. Both imput marged in the same body because at most 1 body parameter can be defined. [Click here for more info](https://github.com/swagger-api/swagger-editor/issues/889)"
        required: true
        schema:
          $ref: "#/definitions/input"

      responses:
        200:
          description: "Filtered Blueprints"
          schema:
            $ref: '#/definitions/output'
  /api/filterBlueprints:
    post:
      summary: "Ranks and filters abstract blueprints based on application requirements"
      description: "The only API of this method filters a list of blueprints (input) that are complaint with the application requirements (input), and ranks them."
      consumes:
      - "application/json; charset=utf-8"
      produces:
      - "application/json; charset=utf-8"
      parameters:
      - in: "body"
        name: "Input"
        description: "Application requirements and list of VDC blueprints. Both imput marged in the same body because at most 1 body parameter can be defined. [Click here for more info](https://github.com/swagger-api/swagger-editor/issues/889)"
        required: true
        schema:
          $ref: "#/definitions/input"

      responses:
        200:
          description: "Filtered Blueprints"
          schema:
            $ref: '#/definitions/output'
      deprecated: true    
definitions:
  applicationRequirements:
    description: "JSON document conforming to application requirements definitions as in https://github.com/DITAS-Project/ApplicationRequirement"
    type: "object"
    example: 
      applicationType: Other
      functionalRequirements:
        methodTags:
          - ask_ICCS_must_match_getAllValuesForBloodTestComponent
        vdcTags:
          - ask_ICCS_must_match_getAllValuesForBloodTestComponent
        schema:
          attributes:
            - SSN
            - component
      attributes:
        dataUtility:
          - id: availability_99
            description: Availability 99
            type: Availability
            properties:
              availability:
                unit: percentage
                minimum: 99
          - id: responseTime_1
            description: ResponseTime 1
            type: ResponseTime
            properties:
              responseTime:
                maximum: 1
                unit: second
          - id: timeliness_06
            description: Timeliness 0.6
            type: Timeliness
            properties:
              timeliness:
                maximum: 0.6
                unit: NONE
        security:
          - id: encryptionAES_128
            description: Encryption AES 128
            type: TransportEncryption
            properties:
              protocol:
                unit: enum
                value: TLS
              algorithm:
                unit: enum
                value: AES
              keyLength:
                unit: number
                minimum: 128
          - id: tracing_1
            description: Tracing
            type: Tracing
            properties:
              level:
                unit: enum
                value: datasource
              sampleRate:
                unit: percentage
                value: 1
              instrumentation:
                unit: enum
                value: VDC
          - id: aclRoleBasedReadOnly
            description: ACL rolebased readOnly
            type: ACL
            properties:
              protocol:
                unit: enum
                value: username/password
              credentials:
                unit: list
                value:
                  - doctor_er_id
        privacy: []
      goalTrees:
        dataUtility:
          type: AND
          children:
            - type: AND
              children:
                - type: OR
                  leaves:
                    - id: serviceAvailable
                      description: Service available
                      weight: 1
                      attributes:
                        - availability_99
                - type: OR
                  leaves:
                    - id: fastDataProcess
                      description: Fast data process
                      weight: 1
                      attributes:
                        - responseTime_1
            - type: AND
              leaves:
                - id: timeliness
                  description: Timeliness
                  weight: 1
                  attributes:
                    - timeliness_06
        security:
          type: AND
          children:
            - type: AND
              leaves:
                - id: storageEncryption
                  description: Storage encryption
                  weight: 1
                  attributes:
                    - encryptionAES_128
          leaves:
            - id: accessControl
              description: Access control
              weight: 1
              attributes:
                - aclRoleBasedReadOnly
            - id: tracing
              description: Tracing and monitoring
              weight: 1
              attributes:
                - tracing_1
...
  candidates:
    properties:
      blueprint:
        $ref: "#/definitions/blueprint"
      methodNames:
        type: "array"
        items:
          type: "string"
          example: "getAllValuesForBloodTestComponent"
  input:
    properties:
      applicationRequirements:
        $ref: "#/definitions/applicationRequirements"
      candidates:
        type: "array"
        items: 
          $ref: "#/definitions/candidates"
  output:
    type: "array"
    items: 
      properties:
        blueprint:
          $ref: "#/definitions/blueprint"
        score:
          type: number
          format: double
          minimum: 0
          maximum: 1
        methodNames:
          type: "array"
          items:
            type: "string"